requirements.txt
Jinja2==3.1.5
SQLAlchemy==2.0.39
streamlit==1.44.1
python-dotenv==1.0.1
pandas==2.2.3
plotly==6.0.1
matplotlib==3.10.1
pytest==8.3.5
pytest-cov==6.0.0
requests==2.32.3
tenacity==9.0.0
ruff==0.11.2